Atualizado classe Categoria, Criado classe CategoriaService e interface CategoriaRepository

+++ src/main/java/com/peojetospring/dominio/Categoria.java
@@ -12,23 +12,25 @@ public class Categoria implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 	
 	public Categoria() {
 		
 	}
+	
 
-	public Categoria(Integer id, String nome) {
+	public Categoria(Long id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Integer getId() {
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -64,7 +66,6 @@ public class Categoria implements Serializable {
 			return false;
 		return true;
 	}
-
